+++ npm/react/examples/craco/craco.config.js
@@ -1,0 +1,16 @@
+const CracoLessPlugin = require('craco-less')
+
+module.exports = {
+  plugins: [
+    {
+      plugin: CracoLessPlugin,
+      options: {
+        lessLoaderOptions: {
+          lessOptions: {
+            javascriptEnabled: true,
+          },
+        },
+      },
+    },
+  ],
+}