+++ packages/reporter/cypress/plugins/index.ts
@@ -1,7 +1,10 @@
-const cp = require('child_process')
+// const cp = require('child_process')
 
-cp.exec('http-server -p 5006 dist')
+// cp.exec('http-server -p 5006 dist')
 
+const express = require('express')
+
+express().use(express.static('dist')).listen(5006)
 /**
  * @type {Cypress.PluginConfig}
  */